feat: Reconcile client secret value
This commit fixes an issue when using server side apply for the Keycloak
client secret k8s secret. In such cases the k8s secret exists, but does not
contains the secret data.

The code added in this commit checks if the existing secret has the client
secret key. If not the client secret key and data will be added to the secret.